requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-allauth==64.2.1
idna==3.8
isodate==0.6.1
lxml==5.3.0
oauthlib==3.2.2
PyJWT==2.9.0
python-dotenv==1.0.1
python3-saml==1.16.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
xmlsec==1.3.14
